moving blog to proprietary
