+++ templates/blog/detail.html
@@ -26,7 +26,7 @@
 {% block content %}
 <article class="bg-white">
   <div class="mx-auto max-w-3xl px-6 py-16">
-    <a href="{% url 'pages:blog_index' %}" class="inline-flex items-center text-sm font-medium text-blue-600 hover:text-blue-700">
+    <a href="{% url 'proprietary:blog_index' %}" class="inline-flex items-center text-sm font-medium text-blue-600 hover:text-blue-700">
       <svg class="mr-2 h-4 w-4" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="1.5" d="M15 19l-7-7 7-7" />
       </svg>
