+++ templates/includes/_footer.html
@@ -16,7 +16,7 @@
       <ul class="space-y-2">
         <li><a href="{% url 'proprietary:about' %}" class="text-slate-700 hover:text-blue-600">About</a></li>
         <li><a href="{% url 'proprietary:careers' %}" class="text-slate-700 hover:text-blue-600">Careers</a></li>
-        <li><a href="{% url 'pages:blog_index' %}" class="text-slate-700 hover:text-blue-600">Blog</a></li>
+        <li><a href="{% url 'proprietary:blog_index' %}" class="text-slate-700 hover:text-blue-600">Blog</a></li>
         <li><a href="mailto:hello@gobii.ai" class="text-slate-700 hover:text-blue-600">Contact</a></li>
       </ul>
     </div>
